0.3.3
Ecco a voi i sorgenti di quella che dovrebbe essere wxBriscola 0.3.3.
Aggiunta la possibilità di cambiare il colore del testo e dello sfondo.
Aggiunta la possibilità di realizzare nuove partite oltre le due mani.
Gli image handerl di wxwidgets ora vengono inizializzati una sola volta e non ogni volta che si crea un nuovo mazzo
Revisione generale del codice per renderlo più leggibile, revisione di tutte le clausole throw
Revisione delle stringhe che devono essere tradotte in ulteriori lingue.

+++ elaboratoreCarteBriscola.cpp
@@ -1,5 +1,5 @@
 /**********************************************************************************
- *   Copyright (C) 2015 by Giulio Sorrentino                                      *
+ *   Copyright (C) 2020 by Giulio Sorrentino                                      *
  *   gsorre84@gmail.com                                                           *
  *                                                                                *
  *   This program is free software; you can redistribute it and/or modify         *
@@ -22,7 +22,7 @@
 
 elaboratoreCarteBriscola::elaboratoreCarteBriscola(bool punti):inizio(true), briscolaDaPunti(punti) {
 	doppione.assign(numeroCarte, false);
-	srand(time(NULL));
+	srand(static_cast<unsigned int>(time(NULL)));
 	cartaBriscola=0;
 }
